misc: change license to from LGPL to MIT
This change of license should not be an issue given that I have been the single
contributor up to this point. The change was made because I do not believe that it is
useful to apply copyleft to something as small as mvui. If mvui ever grows to a size that
would warrant the use of copyleft, then it will have failed as a project in the sense that
the entire point was to build a *minimal* ui framework.